+++ test/encodings/imagepreprocessing.jl
@@ -1,3 +1,5 @@
+include("../imports.jl")
+
 
 @testset "ImagePreprocessing" begin
     encfns = [
@@ -28,9 +30,37 @@
         img = rand(RGB{N0f8}, 10, 10, 10)
         testencoding(enc, block, img)
 
-        enc = ImagePreprocessing(buffered = false)
+        enc = ImagePreprocessing(buffered=false)
         block = Image{3}()
         img = rand(RGB{N0f8}, 10, 10, 10)
         testencoding(enc, block, img)
+    end
+
+    @testset ExtendedTestSet "imagedatasetstats" begin
+
+        @testset ExtendedTestSet "RGB" begin
+            data = [zeros(RGB{Float32}, 10, 10), ones(RGB{Float32}, 10, 10)]
+            means, stds = imagedatasetstats(data, RGB{N0f8}; progress=false)
+            @test means ≈ [0.5, 0.5, 0.5]
+            @test stds ≈ [0., 0., 0.]
+        end
+
+        @testset ExtendedTestSet "Gray" begin
+            data = [zeros(Gray{Float32}, 10, 10), ones(Gray{Float32}, 10, 10)]
+            means, stds = imagedatasetstats(data, Gray{N0f8}; progress=false)
+            @test means ≈ [0.5]
+            @test stds ≈ [0.]
+        end
+
+    end
+
+    @testset ExtendedTestSet "setup" begin
+        data = [
+            zeros(10, 10),
+            ones(10, 10),
+        ]
+        enc = setup(ImagePreprocessing, Image{2}(), data, C = Gray{N0f8})
+        @test enc.stats[1] ≈ [0.5]
+        @test enc.stats[2] ≈ [0.]
     end
 end