+++ test/imports.jl
@@ -3,7 +3,8 @@ using Colors: RGB, N0f8, Gray
 using FastAI
 using FastAI: ParamGroups, IndexGrouper, getgroup, DiscriminativeLRs, decay_optim
 import FastAI: Image, Keypoints, Mask, testencoding, Label, OneHot, ProjectiveTransforms,
-    encodedblock, decodedblock, encode, decode, mockblock, checkblock, Block, Encoding
+    encodedblock, decodedblock, encode, decode, mockblock, checkblock, Block, Encoding,
+    imagedatasetstats
 using FilePathsBase
 using FastAI.Datasets
 using FastAI.Models